Tenant Ten Test/backend/requirements.txt
Django==5.1.6
django-tenants==3.7.0
djangorestframework==3.15.2
psycopg[binary]==3.2.3
python-dotenv==1.0.1
